remove pinia, refactor to use composables instead of plugins, use global middleware, upgrade @azure/msal-browser version
