+++ composables/useAppUser.ts
@@ -1,0 +1,8 @@
+
+
+export const useAppUser = () => {
+    return useState("user", () => ({
+        user: {} as any,
+        userImage: null,
+    }))
+}
