+++ middleware/auth.global.ts
@@ -1,13 +1,15 @@
-import { useUserStore } from "~/stores/auth";
+import {useMSAuth} from "~/composables/useMSAuth";
+import {useAppUser} from "#imports";
 
 export default defineNuxtRouteMiddleware(async (to, from) => {
   if (process.server) return;
+  if (to.name === "/login") return;
 
-  const { $msal } = await useNuxtApp();
-  const accounts = $msal().getAccounts();
-  const userStore = useUserStore();
-  const accessToken = await $msal().acquireTokenSilent();
-  let isAuthenticated = $msal().isAuthenticated() && accessToken;
+  const msAuth = useMSAuth();
+  const accounts = msAuth.getAccounts();
+  const userStore = useAppUser();
+  const accessToken = await msAuth.acquireTokenSilent();
+  let isAuthenticated = msAuth.isAuthenticated() && accessToken;
 
   if (isAuthenticated) {
     const user = {
@@ -16,10 +18,7 @@ export default defineNuxtRouteMiddleware(async (to, from) => {
     };
 
     localStorage.setItem("user", JSON.stringify(user));
-
-    userStore.$patch((state) => {
-      state.user = user;
-    });
+    userStore.value.user = user;
   }
   if (to.name !== "login" && !isAuthenticated) {
     return navigateTo("/login", { replace: true });
