+++ package.json
@@ -20,13 +20,11 @@
     "tailwindcss": "^3.3.2"
   },
   "dependencies": {
-    "@azure/msal-browser": "^3.1.0",
+    "@azure/msal-browser": "^3.5.0",
     "@editorjs/editorjs": "^2.27.0",
     "@editorjs/header": "^2.7.0",
     "@heroicons/vue": "^2.0.18",
     "@nuxt/eslint-config": "^0.1.1",
-    "@pinia/nuxt": "^0.4.11",
-    "eslint": "^8.43.0",
-    "pinia": "^2.1.4"
+    "eslint": "^8.43.0"
   }
 }
