+++ pages/index.vue
@@ -8,7 +8,4 @@
   </div>
 </template>
 <script setup>
-definePageMeta({
-  middleware: ["auth"],
-});
 </script>
