+++ pages/login.vue
@@ -2,8 +2,5 @@
     <LoginScreen />
 </template>
 <script setup>
-definePageMeta({
-    middleware: ["auth"],
-});
 </script>
