--- stores/auth.ts
@@ -1,7 +1,0 @@
-import { defineStore } from "pinia";
-export const useUserStore = defineStore("userdata", {
-  state: () => ({
-    user: {} as any,
-    userImage: null,
-  }),
-});
